Add files via upload
This project explores the journey of Netflix as it emerged as a tech giant in the entertainment industry by dissecting its product and addressing real-world challenges. Through a detailed case study, the paper delves into the problems faced by Netflix and the innovative solutions it employed to overcome them. Additionally, it provides an analysis of Netflix's schema using an entity diagram, shedding light on the intricate workings of its technological infrastructure. This study offers valuable insights into Netflix's evolution and its significant contributions to the entertainment landscape.